1_foundations/community_contributions/SamuelAdebodun/function-app/requirements.txt
azure-functions>=1.18.0,<2.0.0
azure-identity>=1.15.0
azure-storage-blob>=12.19.0
